Added logic for generating snapshot files
* Add All relevant queries and wrappers (to ease testing) to snapshotMainBlockService
* Add test case for NewSnapshotFile and new queries

+++ common/constant/receipt.go
@@ -17,5 +17,5 @@ const (
 	ReceiptGenerateMarkleRootPeriod = 20 * time.Second
 	// number of blocks to lookup for receipts before current height (
 	// this is also the number of blocks to consider when selecting receipts to be included in a snapshot)
-	LinkedReceiptBlocksLimit = 720
+	LinkedReceiptBlocksLimit = uint32(720)
 )
